:recycle: Split send form into token-specific sub-components
