+++ src/components/send/form/formBtc.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import FormBase from './formBase';
+
+export default function FormBtc(props) {
+  return (
+    <FormBase {...props} />
+  );
+}
