+++ src/components/send/form/formLsk.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import FormBase from './formBase';
+
+export default function FormLsk(props) {
+  return (
+    <FormBase {...props} />
+  );
+}
